requirements.txt
presidio-analyzer
presidio-anonymizer
natasha
razdel
pymorphy2
python-docx
pdfminer.six
